Number pad, icons, and not throwing an exception
Added the hdpi, mdpi, xhdpi, and xxhdpi icons and changed the
AndroidManifest.xml file to start showing them in app. They will need to
be changed again in the future so that they look better, but it's a
start.

Also made a small change in activity_main.xml so that only a number pad
shows up for the users, rather than a full keyboard so they are less
likely to break it.

Also changed how it is handled when a user doesn't input anything for
attacker and defender armies. Doesn't throw an exception now, just
checks for length and if it is 0, it returns and shows the toast.